requirements.txt
langchain==0.1.10
openai==1.13.3
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.2